small fix
good progress. but remember we'll need Redis for storing used refresh tokens. otherwise if a token is stolen we can't detect it and close the authentication flow. and to allow the user to have multiple devices connected we need a `clientId` field inside token claims. this will allow us to  recognize which device a user is using so we can only close that specific connection and let others untouched.